+++ Classes/Domain/Repository/SimpleTermRepository.php
@@ -1,0 +1,21 @@
+<?php
+declare(strict_types=1);
+
+namespace Featdd\DpnGlossary\Domain\Repository;
+
+/***
+ *
+ * This file is part of the "dreipunktnull Glossar" Extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ *  (c) 2022 Daniel Dorndorf <dorndorf@featdd.de>
+ *
+ ***/
+
+/**
+ * @package Featdd\DpnGlossary\Domain\Repository
+ */
+class SimpleTermRepository extends AbstractTermRepository
+{}
