+++ Classes/Domain/Repository/TermRepository.php
@@ -15,76 +15,14 @@ namespace Featdd\DpnGlossary\Domain\Repository;
  ***/
 
 use Featdd\DpnGlossary\Domain\Model\Term;
-use TYPO3\CMS\Extbase\Persistence\Exception\InvalidQueryException;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 use TYPO3\CMS\Extbase\Persistence\QueryResultInterface;
-use TYPO3\CMS\Extbase\Persistence\Repository;
 
 /**
  * @package Featdd\DpnGlossary\Domain\Repository
  */
-class TermRepository extends Repository
+class TermRepository extends AbstractTermRepository
 {
-    public const DEFAULT_LIMIT = 5;
-
-    /**
-     * @var array
-     */
-    protected $defaultOrderings = [
-        'name' => QueryInterface::ORDER_ASCENDING,
-    ];
-
-    /**
-     * find all terms sorted by name length
-     *
-     * @return array
-     */
-    public function findByNameLength(): array
-    {
-        $terms = $this->findAll()->toArray();
-
-        /**
-         * Sorting Callback
-         *
-         * @param Term $termA
-         * @param Term $termB
-         * @return int
-         */
-        $sortingCallback = function (Term $termA, Term $termB) {
-            return mb_strlen($termB->getName()) - mb_strlen($termA->getName());
-        };
-
-        // Sort terms
-        usort($terms, $sortingCallback);
-
-        return $terms;
-    }
-
-    /**
-     * finds terms by multiple uids
-     *
-     * @param int[] $uids
-     * @return array
-     */
-    public function findByUids(array $uids): array
-    {
-        $query = $this->createQuery();
-
-        if (0 === count($uids)) {
-            return [];
-        }
-
-        try {
-            $query->matching(
-                $query->in('uid', $uids)
-            );
-        } catch (InvalidQueryException $e) {
-            // nothing
-        }
-
-        return $query->execute()->toArray();
-    }
-
     /**
      * finds the newest terms
      *
