+++ Classes/Service/ParserService.php
@@ -19,8 +19,8 @@ use DOMDocument;
 use DOMNodeList;
 use DOMText;
 use DOMXPath;
-use Featdd\DpnGlossary\Domain\Model\Term;
-use Featdd\DpnGlossary\Domain\Repository\TermRepository;
+use Featdd\DpnGlossary\Domain\Model\SimpleTerm;
+use Featdd\DpnGlossary\Domain\Repository\SimpleTermRepository;
 use Featdd\DpnGlossary\Utility\ParserUtility;
 use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
 use TYPO3\CMS\Core\Context\Context;
@@ -92,8 +92,8 @@ class ParserService implements SingletonInterface
         /** @var \TYPO3\CMS\Extbase\Persistence\Generic\Typo3QuerySettings $querySettings */
         $querySettings = GeneralUtility::makeInstance(Typo3QuerySettings::class);
         // Get termRepository
-        /** @var \Featdd\DpnGlossary\Domain\Repository\TermRepository $termRepository */
-        $termRepository = GeneralUtility::makeInstance(TermRepository::class);
+        /** @var \Featdd\DpnGlossary\Domain\Repository\SimpleTermRepository $simpleTermRepository */
+        $simpleTermRepository = GeneralUtility::makeInstance(SimpleTermRepository::class);
         // Get Typoscript Configuration
         $this->typoScriptConfiguration = $configurationManager->getConfiguration(ConfigurationManagerInterface::CONFIGURATION_TYPE_FULL_TYPOSCRIPT);
         // Reduce TS config to plugin
@@ -130,25 +130,25 @@ class ParserService implements SingletonInterface
             // Set query to respect the language uid
             $querySettings->setRespectSysLanguage(true);
             // Assign query settings object to repository
-            $termRepository->setDefaultQuerySettings($querySettings);
+            $simpleTermRepository->setDefaultQuerySettings($querySettings);
 
             //Find all terms
             if (false === (bool) $this->settings['useCachingFramework']) {
-                $terms = $termRepository->findByNameLength();
+                $terms = $simpleTermRepository->findByNameLength();
             } else {
                 $cacheIdentifier = sha1('termsByNameLength' . $querySettings->getLanguageUid() . '_' . implode('', $querySettings->getStoragePageIds()));
                 $terms = $termsCache->get($cacheIdentifier);
 
                 // If $terms is null, it hasn't been cached. Calculate the value and store it in the cache:
                 if ($terms === false) {
-                    $terms = $termRepository->findByNameLength();
+                    $terms = $simpleTermRepository->findByNameLength();
                     // Save value in cache
                     $termsCache->set($cacheIdentifier, $terms, ['dpnglossary_termscache']);
                 }
             }
 
             //Sort terms with an individual counter for max replacement per page
-            /** @var Term $term */
+            /** @var SimpleTerm $term */
             foreach ($terms as $term) {
                 $maxReplacements = -1 === $term->getMaxReplacements()
                     ? (int) $this->settings['maxReplacementPerPage']
@@ -400,7 +400,7 @@ class ParserService implements SingletonInterface
      */
     protected function domTagsParser(
         DOMNodeList $domTags,
-        Term $term,
+        SimpleTerm $term,
         int &$replacements,
         Closure $wrapperClosure,
         array $forbiddenParentTags
@@ -444,12 +444,12 @@ class ParserService implements SingletonInterface
      * Regex parser for terms on a text string
      *
      * @param string $text
-     * @param Term $term
+     * @param SimpleTerm $term
      * @param int $replacements
      * @param \Closure $wrapperClosure
      * @return string
      */
-    protected function regexParser(string $text, Term $term, int &$replacements, Closure $wrapperClosure): string
+    protected function regexParser(string $text, SimpleTerm $term, int &$replacements, Closure $wrapperClosure): string
     {
         // Try simple search first to save performance
         if (false === mb_stripos($text, $term->getParsingName())) {
@@ -513,11 +513,11 @@ class ParserService implements SingletonInterface
     /**
      * Renders the wrapped term using the plugin settings
      *
-     * @param \Featdd\DpnGlossary\Domain\Model\Term
+     * @param \Featdd\DpnGlossary\Domain\Model\SimpleTerm
      * @return string
      * @throws \UnexpectedValueException
      */
-    protected function termWrapper(Term $term): string
+    protected function termWrapper(SimpleTerm $term): string
     {
         // get content object type
         $contentObjectType = $this->typoScriptConfiguration['settings.']['termWraps'];
@@ -526,7 +526,7 @@ class ParserService implements SingletonInterface
         // pass term data to the cObject pseudo constructor
         $this->contentObjectRenderer->start(
             $term->__toArray(),
-            Term::TABLE
+            SimpleTerm::TABLE
         );
 
         // return the wrapped term
