+++ Configuration/Extbase/Persistence/Classes.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types = 1);
+
+return [
+    \Featdd\DpnGlossary\Domain\Model\SimpleTerm::class => [
+        'tableName' => 'tx_dpnglossary_domain_model_term',
+    ],
+];
